refactor, error handling
